Makefile: Add -Wl,--build-id=none to CFLAGS
Some ARM toolchains emit the .note.gnu.build-id section, which
by default ends up at the start of the binary where the vector
table should be. Use -Wl,--build-id=none in the CFLAGS to suppress
the build-id section entirely so the binary starts with the
vector table as it should.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>