test5: Abort if we failed to get back to privileged mode
This test switches between unprivileged and privileged
modes, which use different stack pointers. If we failed
to get back to privileged mode at the end of the test,
then returning from main() will cause a guest crash
(which will result in an infinite loop of exceptions).
Avoid this by calling abort() if we didn't manage to
get back to the stack pointer we need to safely return
from main().

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>